10.IA/projeto-IA/requirements.txt
flask
requests
python-dotenvn